update to correct type

+++ libs/openFrameworks/math/ofMath.cpp
@@ -12,9 +12,9 @@
 using namespace std;
 
 //--------------------------------------------------
-int ofNextPow2(int a){
+size_t ofNextPow2(size_t a){
 	// from nehe.gamedev.net lesson 43
-	int rval=1;
+	size_t rval=1;
 	while(rval<a) rval<<=1;
 	return rval;
 }
@@ -32,18 +32,18 @@ void ofSeedRandom() {
 		struct timeval tv;
 		gettimeofday(&tv, 0);
 		long int n = (tv.tv_sec ^ tv.tv_usec) ^ getpid();
-		srand(n);
+		srand((unsigned int)n);
 	#else
 		struct timeval tv;
 		gettimeofday(&tv, 0);
 		long int n = (tv.tv_sec ^ tv.tv_usec);
-		srand(n);
+		srand((unsigned int)n);
 	#endif
 }
 
 //--------------------------------------------------
-void ofSeedRandom(int val) {
-	srand((long) val);
+void ofSeedRandom(unsigned int val) {
+	srand((unsigned int)val);
 }
 
 //--------------------------------------------------
